create DesignSystemPreview
