+++ src/routes/AppRoutes.jsx
@@ -1,21 +1,26 @@
-import React from 'react';
-import {Route, Routes} from 'react-router-dom';
-import HomePage from '@pages/home/HomePage';
-import AddRecipePage from '@pages/add-recipe/AddRecipePage';
-import {Layout} from '@components/layout/Layout';
-import RecipeCategories from '@components/recipe-categories/RecipeCategories';
-import RecipeCategoryDetails from '@components/recipe-categories/RecipeCategoryDetails.jsx';
+import React from "react";
+import { Route, Routes } from "react-router-dom";
+import HomePage from "@pages/home/HomePage";
+import AddRecipePage from "@pages/add-recipe/AddRecipePage";
+import DesignSystemPreview from "@pages/design-system/DesignSystemPreview";
+import { Layout } from "@components/layout/Layout";
+import RecipeCategories from "@components/recipe-categories/RecipeCategories";
+import RecipeCategoryDetails from "@components/recipe-categories/RecipeCategoryDetails.jsx";
 
 export const AppRoutes = () => {
-    return (
-        <Routes>
-            <Route element={<Layout/>}>
-                <Route path="/" element={<HomePage/>}>
-                    <Route index element={<RecipeCategories/>}/>
-                    <Route path="category/:categoryId" element={<RecipeCategoryDetails/>}/>
-                </Route>
-                <Route path="/add-recipe" element={<AddRecipePage/>}/>
-            </Route>
-        </Routes>
-    );
+  return (
+    <Routes>
+      <Route element={<Layout />}>
+        <Route path="/" element={<HomePage />}>
+          <Route index element={<RecipeCategories />} />
+          <Route
+            path="category/:categoryId"
+            element={<RecipeCategoryDetails />}
+          />
+        </Route>
+        <Route path="/add-recipe" element={<AddRecipePage />} />
+        <Route path="/design-system" element={<DesignSystemPreview />} />
+      </Route>
+    </Routes>
+  );
 };
